==> analysis/07_analysis_still_to_do.R <==
## Clean up plots directory - try to have all scripts that generate plots as
# numbered scripts inside analysis (see the new files for examples)

## Complete analysis again but this time by spanning over vsly as well by going
## from 0.1 to 2x VSLY and make plot of NMB against VSLY, and policy as different line color
## make these as a facet with respect to vaccine coverage
## this will be figure 4

## also make one by also scaling the assumed cost of a vaccine and plot
## vaccine cost against VSLY scale and do a heatmap of NMB
## make these as a facet with respect to vaccine coverage
## this will be figure 5

## PSA analysis. Sample 100 values based on normal distribution with mean
## as center of value and sd to give 95% CI for each variable we have uncertainty for.
## Then use these values to conduct analysis again with respect to VSLY and vaccine cost
## Again look at difference in facets of different vaccine coverage
## and then do plot of PRCC
